% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLRPM.R
\name{Beta.fun}
\alias{Beta.fun}
\title{Beta function needed in objective function}
\usage{
Beta.fun(a, b)
}
\arguments{
\item{a}{\code{value} specifying Parameter a}

\item{b}{\code{value} specifying Parameter b}
}
\value{
beta returns value of \code{Beta.fun} for parameters a and b
}
\description{
\code{Beta.fun} is a help function for \code{OF}
}
\author{
Christoph Ritschel \email{christoph.ritschel@met.fu-berlin.de}
}

