% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLRPM.R
\docType{data}
\name{BLRPM.class}
\alias{BLRPM.class}
\title{BLRPM class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
BLRPM.class
}
\description{
\code{BLRPM.class} defines a new class for objects of type BLRPM containing the information about storms,
cells, stepfunction and the precipitation time series.
}
\author{
Christoph Ritschel \email{christoph.ritschel@met.fu-berlin.de}
}
\keyword{datasets}

