% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLRPM.R
\name{Delta.fun}
\alias{Delta.fun}
\title{Delta function needed in objective function}
\usage{
Delta.fun(kappa, MStrich)
}
\arguments{
\item{kappa}{\code{value} specifying Parameter kappa}

\item{MStrich}{\code{value} specifying dimension of error correction in objective function}
}
\value{
Delta returns value of \code{Delta.fun} for kappa and MStrich
}
\description{
\code{Delta.fun} is a help function for \code{OF}
}
\author{
Christoph Ritschel \email{christoph.ritschel@met.fu-berlin.de}
}

